backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
argon2-cffi==23.1.0
